+++ database/migrations/2023_08_08_210222_create_gaji_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('set null')->onUpdate('cascade');
             $table->string('nama_karyawan');
             $table->string('jenis_gaji');
-            $table->integer('besar_gaji');
+            $table->bigInteger('besar_gaji');
             $table->timestamps();
         });
     }
