+++ database/migrations/2023_08_22_142635_create_skil_bonus_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('kehadiran_id')->constrained('kehadiran')->onDelete('cascade')->onUpdate('cascade');
             $table->string('nama_bonus');
-            $table->integer('besar_bonus');
+            $table->bigInteger('besar_bonus');
             $table->timestamps();
         });
     }
